pyproject.toml
[project]
name = "titanaboa-essentials"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.2.1",
    "titanoboa>=0.2.8",
]
